put routes in to separate module to fix App intialization in HMR mode #33
