+++ frontend/src/requests/requests.ts
@@ -14,7 +14,7 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-import { router } from "../main.ts";
+import { router } from "../routes.ts";
 
 const host = "api";
